Added JS and Design Tweaks
Changed the layout of the output screen.
Implemented the JavaScript so that calculations now show up when being entered. All buttons now at least add to the current operations screen, Eval doesn't work for divide, multiple or minus, but does for plus and mod. Also need to modify brackets so numbers can be put into them.